refactor: Member에서 auth로 분리
