+++ src/main/java/roomescape/config/WebConfig.java
@@ -6,7 +6,8 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import roomescape.member.LoginMemberArgumentResolver;
+import roomescape.auth.AdminInterceptor;
+import roomescape.auth.LoginMemberArgumentResolver;
 
 @Configuration
 @RequiredArgsConstructor
